refactor: do not split code for no reason, stay close to vscode
